Introduce AbstractTestCase and refactor tests
